changed content to components
