+++ components/About.tsx
@@ -1,0 +1,15 @@
+export default function About() {
+  return (
+    <section id="about" className="mx-auto max-w-7xl px-4 py-16 sm:px-6">
+      <div className="reveal">
+        <h2 className="text-2xl font-semibold">About</h2>
+        <p className="mt-4 max-w-3xl text-gray-200">
+          I'm a hands-on engineer who enjoys solving ambiguous problems. My
+          focus areas include web platforms, data-driven decision making, and
+          designing systems that scale. I collaborate across product, design,
+          and business to turn insights into impact.
+        </p>
+      </div>
+    </section>
+  );
+}
