+++ components/Header.tsx
@@ -1,0 +1,94 @@
+"use client";
+
+import { useEffect, useState } from "react";
+import Link from "next/link";
+import { NAV_LINKS } from "@/constants/navigation";
+
+export default function Header() {
+  const [open, setOpen] = useState(false);
+
+  useEffect(() => {
+    const close = () => setOpen(false);
+    window.addEventListener("hashchange", close);
+    return () => window.removeEventListener("hashchange", close);
+  }, []);
+
+  return (
+    <header className="sticky top-0 z-50 backdrop-blur supports-[backdrop-filter]:bg-[#0D1B2A]/60">
+      <div className="mx-auto max-w-7xl px-4 sm:px-6">
+        <div className="flex h-16 items-center justify-between">
+          <Link href="#home" className="inline-flex items-center gap-2">
+            <span className="text-lg font-semibold tracking-tight">
+              Bimsara Madurapperuma
+            </span>
+            <span className="h-2 w-2 rounded-full bg-[#00BFA6]" aria-hidden />
+          </Link>
+          <nav className="hidden md:block">
+            <ul className="flex items-center gap-6 text-sm">
+              {NAV_LINKS.map((item) => (
+                <li key={item.href}>
+                  <a
+                    href={item.href}
+                    className="link-underline text-gray-200 hover:text-white focus-visible:text-white"
+                  >
+                    {item.label}
+                  </a>
+                </li>
+              ))}
+            </ul>
+          </nav>
+          <button
+            type="button"
+            aria-label="Toggle navigation"
+            onClick={() => setOpen((v) => !v)}
+            className="md:hidden inline-flex h-10 w-10 items-center justify-center rounded border border-white/10 bg-white/5 hover:bg-white/10"
+          >
+            <svg
+              width="20"
+              height="20"
+              viewBox="0 0 24 24"
+              fill="none"
+              className="text-white"
+            >
+              {open ? (
+                <path
+                  d="M6 6l12 12M6 18L18 6"
+                  stroke="currentColor"
+                  strokeWidth="2"
+                  strokeLinecap="round"
+                />
+              ) : (
+                <path
+                  d="M3 6h18M3 12h18M3 18h18"
+                  stroke="currentColor"
+                  strokeWidth="2"
+                  strokeLinecap="round"
+                />
+              )}
+            </svg>
+          </button>
+        </div>
+      </div>
+
+      {open && (
+        <div className="md:hidden border-t border-white/10 bg-[#0D1B2A]/95">
+          <nav className="mx-auto max-w-7xl px-4 sm:px-6">
+            <ul className="flex flex-col py-4 text-base">
+              {NAV_LINKS.map((item) => (
+                <li key={item.href}>
+                  <a
+                    href={item.href}
+                    onClick={() => setOpen(false)}
+                    className="block px-2 py-2 text-gray-200 hover:text-white"
+                  >
+                    {item.label}
+                  </a>
+                </li>
+              ))}
+            </ul>
+          </nav>
+        </div>
+      )}
+    </header>
+  );
+}
