+++ constants/contact.ts
@@ -1,0 +1,9 @@
+export interface ContactLink {
+  label: string;
+  href: string;
+  icon: React.ReactNode;
+}
+
+export const CONTACT_EMAIL = "bimsara@example.com";
+export const LINKEDIN_URL = "https://www.linkedin.com/in/your-profile";
+export const GITHUB_URL = "https://github.com/your-username";
