+++ constants/navigation.ts
@@ -1,0 +1,9 @@
+export const NAV_LINKS = [
+  { href: "#home", label: "Home" },
+  { href: "#about", label: "About" },
+  { href: "#experience", label: "Work" },
+  { href: "#skills", label: "Skills" },
+  { href: "#projects", label: "Projects" },
+  { href: "#awards", label: "Awards" },
+  { href: "#contact", label: "Contact" },
+] as const;
